requirements.txt
crewai>=0.80.0
groq>=0.4.2
python-dotenv>=1.0.0
langchain-groq>=0.2.0
litellm>=1.0.0
